requirements.txt
# Core dependencies
numpy>=1.19.0
requests>=2.25.0
python-dotenv>=0.15.0
pyyaml>=5.4.0
click>=7.1.2

# Web and API
flask>=2.0.0
flask-socketio>=5.1.0
eventlet>=0.30.0
httpx>=0.18.0

# Data processing
pandas>=1.3.0
scikit-learn>=0.24.0

# Utilities
tqdm>=4.61.0
colorama>=0.4.4
python-dateutil>=2.8.1
jsonschema>=3.2.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
